feat: LLMパネルにプロンプトプレビュー機能を追加
