+++ docs/SPECIFICATION.md
@@ -116,6 +116,13 @@
         *   **生成ボタン:**
             *   アイコン: 
             *   機能: プロンプトをLLMに送信し、子ノードを生成
+        *   **プロンプトプレビュー機能:**
+            *   アイコン: 👁️ (目のアイコン)
+            *   機能: LLMに送信されるプロンプトの内容を確認
+            *   表示内容:
+                *   システムプロンプト
+                *   現在のマインドマップの状態
+                *   ユーザーのプロンプト
         *   **生成されたノードのプレビュー:**
             *   テキストエリア: LLMが生成した子ノードをプレビュー表示
         *   **追加ボタン:**
@@ -204,6 +211,7 @@
 *   **作成日:** 2024年5月15日
 *   **作成者:** Gemini
 *   **更新履歴:**
+    *   2025年2月2日: LLMパネルにプロンプトプレビュー機能を追加
     *   2025年2月2日: LLMにツリーの文脈を提供する機能を追加
     *   2025年2月2日: APIルートを廃止し、Server Actionsに完全移行
     *   2025年2月2日: LLMパネルの選択中ノード表示機能を追加
